+++ src/ts/Ads/Ads.ts
@@ -32,7 +32,7 @@ import { AdsConfigurationParser } from 'Ads/Parsers/AdsConfigurationParser';
 import { CustomFeatures } from 'Ads/Utilities/CustomFeatures';
 import { GameSessionCounters } from 'Ads/Utilities/GameSessionCounters';
 import { IosUtils } from 'Ads/Utilities/IosUtils';
-import { ChinaMetric, ProgrammaticTrackingError, MiscellaneousMetric, LoadMetric, TimingMetric, ProgrammaticTrackingService } from 'Ads/Utilities/ProgrammaticTrackingService';
+import { ChinaMetric, ProgrammaticTrackingError, MiscellaneousMetric, LoadMetric, ProgrammaticTrackingService, InitializationMetric } from 'Ads/Utilities/ProgrammaticTrackingService';
 import { SdkStats } from 'Ads/Utilities/SdkStats';
 import { SessionDiagnostics } from 'Ads/Utilities/SessionDiagnostics';
 import { InterstitialWebPlayerContainer } from 'Ads/Utilities/WebPlayer/InterstitialWebPlayerContainer';
@@ -123,6 +123,7 @@ export class Ads implements IAds {
     private _showingPrivacy: boolean = false;
     private _loadApiEnabled: boolean = false;
     private _webViewEnabledLoad: boolean = false;
+    private _mediationName: string;
     private _core: ICore;
 
     public BannerModule: BannerModule;
@@ -181,6 +182,7 @@ export class Ads implements IAds {
         return Promise.resolve().then(() => {
             SdkStats.setInitTimestamp();
             GameSessionCounters.init();
+            Diagnostics.setAbGroup(this._core.Config.getAbGroup());
             return this.setupTestEnvironment();
         }).then(() => {
             return this.configureMediationManager();
@@ -261,7 +263,7 @@ export class Ads implements IAds {
 
             RequestManager.setAuctionProtocol(this._core.Config, this.Config, this._core.NativeBridge.getPlatform(), this._core.ClientInfo);
 
-            this.CampaignManager = new CampaignManager(this._core.NativeBridge.getPlatform(), this._core, this._core.Config, this.Config, this.AssetManager, this.SessionManager, this.AdMobSignalFactory, this._core.RequestManager, this._core.ClientInfo, this._core.DeviceInfo, this._core.MetaDataManager, this._core.CacheBookkeeping, this.ContentTypeHandlerManager, this.PrivacySDK, this.PrivacyManager);
+            this.CampaignManager = new CampaignManager(this._core.NativeBridge.getPlatform(), this._core, this._core.Config, this.Config, this.AssetManager, this.SessionManager, this.AdMobSignalFactory, this._core.RequestManager, this._core.ClientInfo, this._core.DeviceInfo, this._core.MetaDataManager, this._core.CacheBookkeeping, this.ContentTypeHandlerManager, this.PrivacySDK, this.PrivacyManager, this.MediationLoadTrackingManager);
             this.configureRefreshManager();
             SdkStats.initialize(this._core.Api, this._core.RequestManager, this._core.Config, this.Config, this.SessionManager, this.CampaignManager, this._core.MetaDataManager, this._core.ClientInfo, this._core.CacheManager);
 
@@ -276,12 +278,12 @@ export class Ads implements IAds {
         }).then(() => {
             return this._core.Api.Sdk.initComplete();
         }).then(() => {
-            const initializeAuctionTimespan = Date.now();
+            return this.logInitializationLatency();
+        }).then(() => {
             return Promises.voidResult(this.RefreshManager.initialize().then(() => {
                 if (this.MediationLoadTrackingManager) {
                     this.MediationLoadTrackingManager.setInitComplete();
                 }
-                ProgrammaticTrackingService.reportTimingEvent(TimingMetric.AuctionToFillStatusTime, Date.now() - initializeAuctionTimespan);
             }));
         }).then(() => {
             return Promises.voidResult(this.SessionManager.sendUnsentSessions());
@@ -334,12 +336,32 @@ export class Ads implements IAds {
 
             // Potentially use SDK Detection
             return this._core.MetaDataManager.fetch(MediationMetaData).then((mediation) => {
-                if (mediation && mediation.getName()) {
+                if (mediation && mediation.getName() && performance && performance.now) {
+                    this._mediationName = mediation.getName()!;
                     this.MediationLoadTrackingManager = new MediationLoadTrackingManager(this.Api.LoadApi, this.Api.Listener, mediation.getName()!, this._webViewEnabledLoad);
+                    this.MediationLoadTrackingManager.reportPlacementCount(this.Config.getPlacementCount());
                 }
             }).catch();
         }
         return Promise.resolve();
+    }
+
+    private logInitializationLatency(): void {
+        //tslint:disable-next-line
+        const initTimestamp = (<any>window).unityAdsWebviewInitTimestamp;
+        if (initTimestamp && performance && performance.now) {
+            const webviewInitTime = performance.now() - initTimestamp;
+            const tags = [
+                ProgrammaticTrackingService.createAdsSdkTag('wel', `${this._webViewEnabledLoad}`),
+                ProgrammaticTrackingService.createAdsSdkTag('lae', `${this._loadApiEnabled}`)
+            ];
+
+            if (this._mediationName) {
+                tags.push(ProgrammaticTrackingService.createAdsSdkTag('med', this._mediationName));
+            }
+            ProgrammaticTrackingService.reportTimingEventWithTags(InitializationMetric.WebviewInitialization, webviewInitTime, tags);
+            ProgrammaticTrackingService.reportTimingEventWithTags(InitializationMetric.WebviewPageLoading, initTimestamp, tags);
+        }
     }
 
     private showPrivacyIfNeeded(options: unknown): Promise<void> {