+++ src/ts/Ads/Networking/MetricInstance.spec.ts
@@ -1,9 +1,8 @@
 import { RequestManager, RequestManagerMock } from 'Core/Managers/__mocks__/RequestManager';
 import { ClientInfo, ClientInfoMock } from 'Core/Models/__mocks__/ClientInfo';
 import { DeviceInfo, DeviceInfoMock } from 'Core/Models/__mocks__/DeviceInfo';
-
 import { IProgrammaticTrackingData, MetricInstance } from 'Ads/Networking/MetricInstance';
-import { AdmobMetric, ProgrammaticTrackingError, TimingMetric, BannerMetric } from 'Ads/Utilities/ProgrammaticTrackingService';
+import { AdmobMetric, ProgrammaticTrackingError, TimingEvent, InitializationMetric, MediationMetric, BannerMetric } from 'Ads/Utilities/ProgrammaticTrackingService';
 import { Platform } from 'Core/Constants/Platform';
 
 [
@@ -16,7 +15,7 @@ import { Platform } from 'Core/Constants/Platform';
     let requestManager: RequestManagerMock;
     const osVersion = '11.2.1';
     const sdkVersion = '2300';
-    const country = 'us';
+    const country = 'US';
 
     let metricInstance: MetricInstance;
 
@@ -116,7 +115,7 @@ import { Platform } from 'Core/Constants/Platform';
                     'https://sdk-diagnostics.prd.mz.internal.unity3d.com/v1/metrics',
                     JSON.stringify(t.expected),
                     [['Content-Type', 'application/json']]
-                    );
+                );
 
                 return promise;
             });
@@ -135,8 +134,8 @@ import { Platform } from 'Core/Constants/Platform';
                         name: 'admob_used_cached_video',
                         value: 1,
                         tags: [
-                            'ads_sdk2_mevt:admob_used_cached_video',
                             `ads_sdk2_sdv:${sdkVersion}`,
+                            'ads_sdk2_iso:us',
                             `ads_sdk2_plt:${Platform[platform]}`
                         ]
                     }
@@ -150,8 +149,8 @@ import { Platform } from 'Core/Constants/Platform';
                         name: 'admob_used_streamed_video',
                         value: 1,
                         tags: [
-                            'ads_sdk2_mevt:admob_used_streamed_video',
                             `ads_sdk2_sdv:${sdkVersion}`,
+                            'ads_sdk2_iso:us',
                             `ads_sdk2_plt:${Platform[platform]}`
                         ]
                     }
@@ -177,7 +176,7 @@ import { Platform } from 'Core/Constants/Platform';
                     'https://sdk-diagnostics.prd.mz.internal.unity3d.com/v1/metrics',
                     JSON.stringify(t.expected),
                     [['Content-Type', 'application/json']]
-                    );
+                );
 
                 return promise;
             });
@@ -199,8 +198,8 @@ import { Platform } from 'Core/Constants/Platform';
                         name: 'admob_used_cached_video',
                         value: 1,
                         tags: [
-                            'ads_sdk2_mevt:admob_used_cached_video',
                             `ads_sdk2_sdv:${sdkVersion}`,
+                            'ads_sdk2_iso:us',
                             `ads_sdk2_plt:${Platform[platform]}`,
                             'ads_sdk2_blt:3.0.0'
                         ]
@@ -216,8 +215,8 @@ import { Platform } from 'Core/Constants/Platform';
                         name: 'admob_used_streamed_video',
                         value: 1,
                         tags: [
-                            'ads_sdk2_mevt:admob_used_streamed_video',
                             `ads_sdk2_sdv:${sdkVersion}`,
+                            'ads_sdk2_iso:us',
                             `ads_sdk2_plt:${Platform[platform]}`,
                             'ads_sdk2_test:testValue'
                         ]
@@ -244,7 +243,7 @@ import { Platform } from 'Core/Constants/Platform';
                     'https://sdk-diagnostics.prd.mz.internal.unity3d.com/v1/metrics',
                     JSON.stringify(t.expected),
                     [['Content-Type', 'application/json']]
-                    );
+                );
 
                 return promise;
             });
@@ -254,12 +253,12 @@ import { Platform } from 'Core/Constants/Platform';
     describe('reportTimingEvent', () => {
 
         const tests: {
-            metric: TimingMetric;
+            metric: TimingEvent;
             value: number;
             path: string;
             expected: IProgrammaticTrackingData;
         }[] = [{
-            metric: TimingMetric.TotalWebviewInitializationTime,
+            metric: InitializationMetric.WebviewInitialization,
             value: 18331,
             path: '/timing',
             expected: {
@@ -276,7 +275,7 @@ import { Platform } from 'Core/Constants/Platform';
                 ]
             }
         }, {
-            metric: TimingMetric.TotalWebviewInitializationTime,
+            metric: InitializationMetric.WebviewInitialization,
             value: -1,
             path: '/metrics',
             expected: {
@@ -285,9 +284,10 @@ import { Platform } from 'Core/Constants/Platform';
                         name: 'timing_value_negative',
                         value: 1,
                         tags: [
-                            'ads_sdk2_mevt:webview_initialization_time', // Intentional to track which timing metrics are negative
                             `ads_sdk2_sdv:${sdkVersion}`,
-                            `ads_sdk2_plt:${Platform[platform]}`
+                            'ads_sdk2_iso:us',
+                            `ads_sdk2_plt:${Platform[platform]}`,
+                            'ads_sdk2_mevt:webview_initialization_time' // Intentional to track which timing metrics are negative
                         ]
                     }
                 ]
@@ -312,9 +312,40 @@ import { Platform } from 'Core/Constants/Platform';
                     'https://sdk-diagnostics.prd.mz.internal.unity3d.com/v1' + t.path,
                     JSON.stringify(t.expected),
                     [['Content-Type', 'application/json']]
-                    );
+                );
 
                 return promise;
+            });
+        });
+
+        describe('with untracked countries', () => {
+            beforeEach(() => {
+                metricInstance = new MetricInstance(platform, requestManager, clientInfo, deviceInfo, 'mz');
+                metricInstance.reportTimingEvent(InitializationMetric.WebviewInitialization, 10);
+                return metricInstance.sendBatchedEvents();
+            });
+
+            it('should call with the expected row country', () => {
+                const expected = {
+                    metrics: [
+                        {
+                            name: 'webview_initialization_time',
+                            value: 10,
+                            tags: [
+                                `ads_sdk2_sdv:${sdkVersion}`,
+                                'ads_sdk2_iso:row',
+                                `ads_sdk2_plt:${Platform[platform]}`
+                            ]
+                        }
+                    ]
+                };
+
+                expect(requestManager.post).toBeCalledWith(
+                    'https://sdk-diagnostics.prd.mz.internal.unity3d.com/v1/timing',
+                    JSON.stringify(expected),
+                    [['Content-Type', 'application/json']]
+                );
+
             });
         });
     });
@@ -327,18 +358,11 @@ import { Platform } from 'Core/Constants/Platform';
             });
         });
 
-        it('should not fire events when negative valued events are batched', () => {
-            metricInstance.reportTimingEvent(TimingMetric.AdsInitializeTime, -200);
-            return metricInstance.sendBatchedTimingEvents().then(() => {
-                expect(requestManager.post).toBeCalledTimes(0);
-            });
-        });
-
         describe('Batch two timing events', () => {
 
             beforeEach(() => {
-                metricInstance.reportTimingEvent(TimingMetric.CoreInitializeTime, 999);
-                metricInstance.reportTimingEvent(TimingMetric.WebviewLoadToConfigurationCompleteTime, 100);
+                metricInstance.reportTimingEvent(InitializationMetric.WebviewInitialization, 999);
+                metricInstance.reportTimingEvent(MediationMetric.LoadRequestNofill, 100);
             });
 
             it('should call post once', () => {
@@ -351,7 +375,7 @@ import { Platform } from 'Core/Constants/Platform';
                 const expected = {
                     metrics: [
                         {
-                            name: 'uads_core_initialize_time',
+                            name: 'webview_initialization_time',
                             value: 999,
                             tags: [
                                 `ads_sdk2_sdv:${sdkVersion}`,
@@ -359,7 +383,7 @@ import { Platform } from 'Core/Constants/Platform';
                                 `ads_sdk2_plt:${Platform[platform]}`
                             ]
                         }, {
-                            name: 'webview_load_to_configuration_complete_time',
+                            name: 'load_request_nofill_time',
                             value: 100,
                             tags: [
                                 `ads_sdk2_sdv:${sdkVersion}`,
@@ -374,7 +398,7 @@ import { Platform } from 'Core/Constants/Platform';
                     'https://sdk-diagnostics.prd.mz.internal.unity3d.com/v1/timing',
                     JSON.stringify(expected),
                     [['Content-Type', 'application/json']]
-                    );
+                );
 
                 return promise;
             });
@@ -406,16 +430,16 @@ import { Platform } from 'Core/Constants/Platform';
                             name: 'banner_ad_load',
                             value: 1,
                             tags: [
-                                'ads_sdk2_mevt:banner_ad_load',
                                 `ads_sdk2_sdv:${sdkVersion}`,
+                                'ads_sdk2_iso:us',
                                 `ads_sdk2_plt:${Platform[platform]}`
                             ]
                         }, {
                             name: 'banner_ad_impression',
                             value: 1,
                             tags: [
-                                'ads_sdk2_mevt:banner_ad_impression',
                                 `ads_sdk2_sdv:${sdkVersion}`,
+                                'ads_sdk2_iso:us',
                                 `ads_sdk2_plt:${Platform[platform]}`
                             ]
                         }
@@ -426,7 +450,7 @@ import { Platform } from 'Core/Constants/Platform';
                     'https://sdk-diagnostics.prd.mz.internal.unity3d.com/v1/metrics',
                     JSON.stringify(expected),
                     [['Content-Type', 'application/json']]
-                    );
+                );
 
                 return promise;
             });
@@ -443,13 +467,13 @@ import { Platform } from 'Core/Constants/Platform';
             it('should not fire events when below 10', () => {
                 for (let i = 0; i < 10; i++) {
                     expect(requestManager.post).toBeCalledTimes(0);
-                    metricInstance.reportTimingEvent(TimingMetric.TotalWebviewInitializationTime, 200);
+                    metricInstance.reportTimingEvent(InitializationMetric.WebviewInitialization, 200);
                 }
             });
 
             it('should fire events when 10 events are reached', () => {
                 for (let i = 0; i < 10; i++) {
-                    metricInstance.reportTimingEvent(TimingMetric.TotalWebviewInitializationTime, 200);
+                    metricInstance.reportTimingEvent(InitializationMetric.WebviewInitialization, 200);
                 }
                 expect(requestManager.post).toBeCalledTimes(1);
             });
@@ -460,7 +484,7 @@ import { Platform } from 'Core/Constants/Platform';
         beforeEach(() => {
             clientInfo.getTestMode.mockReturnValue(true);
             metricInstance = new MetricInstance(platform, requestManager, clientInfo, deviceInfo, country);
-            for (let x=0;x<11;x++) {
+            for (let i = 0; i < 11; i++) {
                 metricInstance.reportMetricEvent(AdmobMetric.AdmobUsedStreamedVideo);
             }
         });
@@ -473,5 +497,4 @@ import { Platform } from 'Core/Constants/Platform';
             );
         });
     });
-
 }));